backend/requirements.txt
# === Web Framework ===
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
jinja2>=3.1.3

# === LLM Providers ===
openai>=1.12.0
anthropic>=0.18.0
google-genai>=1.0.0
httpx>=0.27.0

# === Vector Store ===
chromadb>=0.4.22

# === Document Processing ===
pymupdf>=1.23.0
python-docx>=1.1.0
openpyxl>=3.1.0
python-pptx>=0.6.23

# === Configuration ===
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
PyYAML>=6.0.1

# === Utilities ===
aiofiles>=23.2.0
python-jose[cryptography]>=3.3.0

# === Auth ===
bcrypt>=4.0.0
